+++ ui/src/core/components/PromptInput/PromptInputText.tsx
@@ -1,25 +1,14 @@
 import type { TextAreaProps } from "@heroui/react";
-
-import React from "react";
+import { forwardRef } from "react";
 import { Textarea } from "@heroui/react";
-import { cn } from "@heroui/react";
 
-const PromptInputText = React.forwardRef<HTMLTextAreaElement, TextAreaProps>(
+const PromptInputText = forwardRef<HTMLTextAreaElement, TextAreaProps>(
   ({ classNames = {}, ...props }, ref) => {
     return (
       <Textarea
         ref={ref}
-        aria-label="Prompt"
-        className="min-h-[40px]"
-        classNames={{
-          ...classNames,
-          label: cn("hidden", classNames?.label),
-          input: cn("py-0", classNames?.input),
-        }}
-        minRows={1}
-        placeholder="Enter a prompt here"
-        radius="lg"
-        variant="bordered"
+        className="min-h-[50px]"
+        classNames={classNames}
         {...props}
       />
     );
@@ -27,5 +16,3 @@ const PromptInputText = React.forwardRef<HTMLTextAreaElement, TextAreaProps>(
 );
 
 export default PromptInputText;
-
-PromptInputText.displayName = "PromptInputText";