+++ packages/ragbits-chat/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Add uvicorn to dependencies (#578)
+- Remove HeroUI Pro components (#557)
+
 ## 0.18.0 (2025-05-22)
 
 ### Changed